UPDATE: plotNeutrophilMovie.m different data input.
Added functionality to use the function plotNeutrophilMovie, without
depending on the _mat_Re/ and _mat_La/ directories. User is able to
search for a folder which contains different data from the one processed
by phagosight.

Usage:

>> plotNeutrophilMovie(handlesDir, newDataDir)

where handlesDir, as before, is the folder that contains the handles
structure. The variable newDataDir cna be one of two options:

(1) The path to folder with data (.mat and .tif files supported)
(2) 'The string folder', which allow the user to select the folder to
work with.